requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
waitress==2.1.2
python-dotenv==1.0.0
